+++ pcsx2/VU1micro.cpp
@@ -31,7 +31,7 @@ void vu1Finish(bool add_cycles) {
 		vu1Thread.Get_MTVUChanges();
 		return;
 	}
-	u32 vu1cycles = VU1.cycle;
+	u64 vu1cycles = VU1.cycle;
 	if(VU0.VI[REG_VPU_STAT].UL & 0x100) {
 		VUM_LOG("vu1ExecMicro > Stalling until current microprogram finishes");
 		CpuVU1->Execute(vu1RunCycles);
